refactor: 데이터로직과 ui로직 분리
