+++ src/components/hook/useReturnVisitRate.ts
@@ -1,0 +1,14 @@
+import { fetchData } from '@/utils/api';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+interface ReturnVisitRateDataType {
+  domain: string;
+  visitedUsersRate: number;
+}
+
+export const useReturnVisitRate = () => {
+  return useSuspenseQuery<ReturnVisitRateDataType>({
+    queryKey: ['returnVisitRate'],
+    queryFn: () => fetchData('/dashboard/visitedRate'),
+  });
+};
