+++ src/components/hook/useTotalVisitors.ts
@@ -1,0 +1,14 @@
+import { fetchData } from '@/utils/api';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+interface TotalVisitorsDataType {
+  totalVisitCount: string;
+  uniqueVisitors: number;
+}
+
+export const useTotalVisitors = () => {
+  return useSuspenseQuery<TotalVisitorsDataType[]>({
+    queryKey: ['totalVisitors'],
+    queryFn: () => fetchData('/dashboard/totalVisitorsCount'),
+  });
+};
