+++ src/components/hook/useUserLanguages.ts
@@ -1,0 +1,14 @@
+import { fetchData } from '@/utils/api';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+interface UserLanguagesDataType {
+  language: string;
+  count: number;
+}
+
+export const useUserLanguages = () => {
+  return useSuspenseQuery<UserLanguagesDataType[]>({
+    queryKey: ['userLanguage'],
+    queryFn: () => fetchData('/dashboard/languageStats'),
+  });
+};
