+++ src/components/stats/UserLanguages.tsx
@@ -1,26 +1,17 @@
-import { fetchData } from '@/utils/api';
-import { useSuspenseQuery } from '@tanstack/react-query';
 import { PieChartTemplate } from '../charts/PieChartTemplate';
 import { Card } from '../common/Card';
 import { Title } from '../common/Title';
-
-interface UserLanguagesDataType {
-  language: string;
-  count: number;
-}
+import { useUserLanguages } from '../hook/useUserLanguages';
 
 export const UserLanguages = () => {
-  const { data } = useSuspenseQuery<UserLanguagesDataType[]>({
-    queryKey: ['userLanguage'],
-    queryFn: () => fetchData('/dashboard/languageStats'),
-  });
-
+  const { data } = useUserLanguages();
   const languageData = data.map(
     (item: { language: string; count: number }) => ({
       name: item.language,
       value: item.count,
     })
   );
+
   return (
     <Card width="fit">
       <Title title="Language Stats" />
